pyproject.toml
[project]
name = "resume-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "pdfplumber>=0.11.9",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_scanner"]
